requirements.txt
Django==4.2
asgiref==3.7.2
attrs==23.1.0
beautifulsoup4==4.12.2
boto3==1.28.28
botocore==1.31.28
certifi==2023.7.22
charset-normalizer==3.2.0
colorama==0.4.6
distlib==0.3.7
django-cors-headers==4.2.0
django-environ==0.10.0
django-jsonfield==1.4.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
exceptiongroup==1.1.1
filelock==3.12.2
flake8==6.0.0
gunicorn==21.2.0
idna==3.4
iniconfig==2.0.0
jmespath==1.0.1
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
mccabe==0.7.0
numpy==1.24.3
packaging==23.1
pandas==2.0.1
Pillow==10.0.0
platformdirs==3.10.0
pluggy==1.0.0
psycopg==3.1.10
psycopg2==2.9.7
psycopg2-binary==2.9.7
pycodestyle==2.10.0
pyflakes==3.0.1
PyJWT==2.8.0
pyodbc==4.0.39
pyrsistent==0.19.3
pytest==7.3.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-slugify==8.0.1
pytz==2023.3
referencing==0.30.2
regex==2023.8.8
requests==2.31.0
rpds-py==0.9.2
s3transfer==0.6.2
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.4
text-unidecode==1.3
tomli==2.0.1
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.16
uuid==1.30
virtualenv==20.24.3